Updating API to reflect latest changes (#244)
* chore: bump taikungoclient and remove deprecations

Remove:
* project user binding
* user tokens
* replace orgs by account

* Small refactoring of a Makefile

* Fixing build

* Fixing some tests

* Small refactor

* Fix dep

* Fixing more unit tests

* Fixing repository enable command

* Cloudcredential UT fix

* Fixing user (redundand endpoint)

* Implementing group API

* Adding group UTs

* Adding unit tests for groups/organizations and groups/users endpoints

* Linking go.mod to latest taikungoclient on 'app-latest'

* Completing groups API endpoint implementation

* Fixing showback implementation

* Fixing build for goreleaser

* Adding robot user endpoint

* Addin accounts endpoint

* Bumping taikungoclient dependency to main branch

* Removing unnecessary unit tests

* hotfix: Minimum stuff to solve https://cloudera.atlassian.net/browse/CCF-218
 and unblock other teams. AWC CC had new DTO and CLI failed silently.

* Bumping taikungoclient in go.mod

* Adding subnet listing within a project

* Fixing accounts endpoint

* Little rework of organization

* Extending cloud credentials and user

* Fixing robot endpoint

* Updating go.mod, fixing linters, and excluding user projects list command

---------

+++ cmd/billing/rule/list/list.go
@@ -92,8 +92,8 @@ func listRun(opts *ListOptions) (err error) {
 
 		billingRules = append(billingRules, data.GetData()...)
 
-		count := int32(len(billingRules))
-		if opts.Limit != 0 && count >= opts.Limit {
+		count := int64(len(billingRules))
+		if opts.Limit != 0 && count >= int64(opts.Limit) {
 			break
 		}
 
@@ -101,7 +101,7 @@ func listRun(opts *ListOptions) (err error) {
 			break
 		}
 
-		myRequest = myRequest.Offset(count)
+		myRequest = myRequest.Offset(int32(count))
 	}
 
 	if opts.Limit != 0 && int32(len(billingRules)) > opts.Limit {
