+++ cmd/cmdutils/flavors_completion.go
@@ -3,6 +3,7 @@ package cmdutils
 import (
 	"context"
 	"fmt"
+
 	"github.com/itera-io/taikun-cli/utils/types"
 	tk "github.com/itera-io/taikungoclient"
 	"github.com/spf13/cobra"
@@ -32,13 +33,13 @@ func FlavorCompletionFunc(cmd *cobra.Command, args []string, toComplete string) 
 			completions = append(completions, flavor.GetName())
 		}
 
-		count := int32(len(completions))
+		count := int64(len(completions))
 
 		if count == data.GetTotalCount() {
 			break
 		}
 
-		myRequest = myRequest.Offset(count)
+		myRequest = myRequest.Offset(int32(count))
 	}
 
 	return completions
